+++ gomap/map_test.go
@@ -1,8 +1,6 @@
 package gomap_test
 
 import (
-	"strconv"
-	"sync"
 	"testing"
 
 	"github.com/philippgille/gokv/gomap"
@@ -47,29 +45,7 @@ func TestStoreConcurrent(t *testing.T) {
 
 	goroutineCount := 1000
 
-	waitGroup := sync.WaitGroup{}
-	waitGroup.Add(goroutineCount) // Must be called before any goroutine is started
-	for i := 0; i < goroutineCount; i++ {
-		go test.InteractWithStore(store, strconv.Itoa(i), t, &waitGroup)
-	}
-	waitGroup.Wait()
-
-	// Now make sure that all values are in the store
-	expected := test.Foo{}
-	for i := 0; i < goroutineCount; i++ {
-		actualPtr := new(test.Foo)
-		found, err := store.Get(strconv.Itoa(i), actualPtr)
-		if err != nil {
-			t.Errorf("An error occurred during the test: %v", err)
-		}
-		if !found {
-			t.Error("No value was found, but should have been")
-		}
-		actual := *actualPtr
-		if actual != expected {
-			t.Errorf("Expected: %v, but was: %v", expected, actual)
-		}
-	}
+	test.TestConcurrentInteractions(t, goroutineCount, store)
 }
 
 // TestErrors tests some error cases.