+++ memcached/memcached_test.go
@@ -2,8 +2,6 @@ package memcached_test
 
 import (
 	"log"
-	"strconv"
-	"sync"
 	"testing"
 	"time"
 
@@ -70,29 +68,7 @@ func TestClientConcurrent(t *testing.T) {
 	// Looks like the server load is too high, but should that really be the case with Memcached?
 	goroutineCount := 250
 
-	waitGroup := sync.WaitGroup{}
-	waitGroup.Add(goroutineCount) // Must be called before any goroutine is started
-	for i := 0; i < goroutineCount; i++ {
-		go test.InteractWithStore(client, strconv.Itoa(i), t, &waitGroup)
-	}
-	waitGroup.Wait()
-
-	// Now make sure that all values are in the store
-	expected := test.Foo{}
-	for i := 0; i < goroutineCount; i++ {
-		actualPtr := new(test.Foo)
-		found, err := client.Get(strconv.Itoa(i), actualPtr)
-		if err != nil {
-			t.Errorf("An error occurred during the test: %v", err)
-		}
-		if !found {
-			t.Error("No value was found, but should have been")
-		}
-		actual := *actualPtr
-		if actual != expected {
-			t.Errorf("Expected: %v, but was: %v", expected, actual)
-		}
-	}
+	test.TestConcurrentInteractions(t, goroutineCount, client)
 }
 
 // TestErrors tests some error cases.