+++ test/test.go
@@ -264,6 +264,33 @@ func handleGetError(t *testing.T, err error, found bool) {
 	}
 }
 
+// TestConcurrentInteractions launches a bunch of goroutines that concurrently work with the store.
+func TestConcurrentInteractions(t *testing.T, goroutineCount int, store gokv.Store) {
+	waitGroup := sync.WaitGroup{}
+	waitGroup.Add(goroutineCount) // Must be called before any goroutine is started
+	for i := 0; i < goroutineCount; i++ {
+		go InteractWithStore(store, strconv.Itoa(i), t, &waitGroup)
+	}
+	waitGroup.Wait()
+
+	// Now make sure that all values are in the store
+	expected := Foo{}
+	for i := 0; i < goroutineCount; i++ {
+		actualPtr := new(Foo)
+		found, err := store.Get(strconv.Itoa(i), actualPtr)
+		if err != nil {
+			t.Errorf("An error occurred during the test: %v", err)
+		}
+		if !found {
+			t.Error("No value was found, but should have been")
+		}
+		actual := *actualPtr
+		if actual != expected {
+			t.Errorf("Expected: %v, but was: %v", expected, actual)
+		}
+	}
+}
+
 // InteractWithStore reads from and writes to the DB. Meant to be executed in a goroutine.
 // Does NOT check if the DB works correctly (that's done elsewhere),
 // only checks for errors that might occur due to concurrent access.