+++ RELEASES.md
@@ -8,12 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 vNext
 -----
 
-- Added: method `Close() error` (issue [#36](https://github.com/philippgille/gokv/issues/36))
+- Added: Method `Close() error` (issue [#36](https://github.com/philippgille/gokv/issues/36))
 - Added: Package `mongodb` - A `gokv.Store` implementation for [MongoDB](https://github.com/mongodb/mongo) (issue [#27](https://github.com/philippgille/gokv/issues/27))
 - Added: Package `dynamodb` - A `gokv.Store` implementation for [Amazon DynamoDB](https://aws.amazon.com/dynamodb/) (issue [#28](https://github.com/philippgille/gokv/issues/28))
 - Added: Package `memcached` - A `gokv.Store` implementation for [Memcached](https://github.com/memcached/memcached) (issue [#31](https://github.com/philippgille/gokv/issues/31))
 - Added: Package `mysql` - A `gokv.Store` implementation for [MySQL](https://github.com/mysql/mysql-server) (issue [#32](https://github.com/philippgille/gokv/issues/32))
 - Added: The factory function `redis.NewClient()` now checks if the connection to the Redis server works and otherwise returns an error.
+- Added: The `test` package now has the function `func TestConcurrentInteractions(t *testing.T, goroutineCount int, store gokv.Store)` that you can use to test your `gokv.Store` implementation with concurrent interactions.
 - Fixed spelling in error message when using the etcd implementation and the etcd server is unreachable
 
 ### Breaking changes